refactor by parent class
