+++ recommender/model/recommender_base.py
@@ -2,21 +2,10 @@ from abc import abstractmethod
 from tools.topk import topk
 
 
-class ImplicitMatrixFactorizationBase:
+class RecommenderBase:
     def __init__(self):
         self.user_factors = None
         self.item_factors = None
-
-    @abstractmethod
-    def fit(self, user_items, val_user_items):
-        """
-        Trains recommendation algorithm on a sparse matrix
-
-        Parameters
-        ----------
-        user_items : csr_matrix
-        """
-        raise NotImplementedError
 
     def recommend(self, userid, user_items, filter_already_liked_items=True, N=10):
         """
@@ -58,22 +47,5 @@ class ImplicitMatrixFactorizationBase:
         Returns
         -------
         user_item_score: M1 x N (np.ndarray)
-        """
-        raise NotImplementedError
-
-    @abstractmethod
-    def calculate_loss(self, user_items):
-        """
-        Calculates training / validation loss for early stopping to prevent overfitting.
-        This function is run after fitting user_factors and item_factors.
-
-        Parameters
-        ----------
-        user_items : M x N user item matrix (csr)
-            Dataset could be train set or validation set
-
-        Returns
-        -------
-        loss
         """
         raise NotImplementedError
